refactor: 검증 로직 위치 변경

+++ src/main/java/roomescape/application/ReservationService.java
@@ -7,6 +7,7 @@ import roomescape.application.dto.ReservationDto;
 import roomescape.application.dto.ReservationInfoDto;
 import roomescape.domain.Reservation;
 import roomescape.domain.repository.ReservationRepository;
+import roomescape.exception.NotFoundReservationException;
 
 import java.util.List;
 
@@ -33,6 +34,9 @@ public class ReservationService {
     }
 
     public int delete(Long id) {
+        if (reservationRepository.count() == 0) {
+            throw new NotFoundReservationException("존재하지 않는 예약입니다.");
+        }
         return reservationRepository.delete(id);
     }
 }
